refactor: CategoryModel 위치 이동
